error in subsection_details.html
the SME is working, but the web page is having an error

ValueError at /online_help/subsection_details/23/
Field 'id' expected a number but got 'Verna Deatras'.

Still cannot add writer.
additionally, in the tasks_test, only object is shows for the SME column, not the actual name (string) of the SME

I pushed so that we have a backup in the repo