Adding writer and SME are working (with bugs)
problem 1: when a particular task is unassigned, and you clicked the subsection of that task, it will show the writer blank with the remove button beside.

Once you remove that empty writer, the task would be gone. I suppose that a task would not be in database if it hasn't contain any writer (even a blank writer).

Problem 2: how can we add a blank/unassigned writer? because if we don't add a blank writer, the task would not created (kahit gumawa ka ng docu, sect, and subsect)

Problem 3: duplicated SMEs

Problem 4: Edit button and text field of SME are not placed well