fix the dash_per_user
We have the draft for dash_per_user now (previously known as per_user_test). But we got an error when clicking the subsection name

MultipleObjectsReturned at /online_help/view_subsection/157/
get() returned more than one Task -- it returned 4!

This happens to per_subsection click in the home_test when multiple IDs are being pass. For example, 'Accessing Command Help and Command Options in the Tcl Console' has four writers. Then when you pass it in the url, apat rin ang pinapasa niya, making the error.

I push this so that we will have a backup

fixed:
-draft of dash_per_user
-navbar for 'All'
-download excel is now functioning

to do:
-fix the dash_per_user